requirements.txt
APScheduler==3.3.1
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
backports.weakref==1.0rc1
bitarray==0.8.1
bleach==1.5.0
boto==2.48.0
bz2file==0.98
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
Cython==0.26
enum34==1.1.6
funcsigs==1.0.2
futures==3.1.1
gensim==2.3.0
html5lib==0.9999999
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
jieba==0.38
lxml==3.8.0
Markdown==2.6.8
mock==2.0.0
MySQL-python==1.2.5
numpy==1.13.1
olefile==0.44
parsel==1.2.0
pbr==3.1.1
Pillow==4.2.1
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pybloom==1.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.1.0
python-dateutil==2.6.1
pytz==2017.2
queuelib==1.4.2
requests==2.18.1
scipy==0.19.1
Scrapy==1.4.0
service-identity==17.0.0
singleton==0.1.0
six==1.10.0
smart-open==1.5.3
SQLAlchemy==1.1.11
tensorflow==1.2.1
tflearn==0.3.2
Twisted==17.5.0
tzlocal==1.4
urllib3==1.21.1
w3lib==1.17.0
Werkzeug==0.12.2
word2vec==0.9.1
zope.interface==4.4.2
